requirements/requirements-gpu.txt
tqdm==4.15.0
pandas==0.22.0
tflearn==0.3.2
numpy==1.14.2
tensorflow-gpu==1.10.0
gensim==3.4.0